refactor(compose): the voice half of reply drafting is its own file
replydraft.go crossed the 500-line cap in #952 (448 → 539), which turns
deterministic-gates red on main and therefore on every PR, since a PR is tested
merged with main. Closes #953.

Split rather than waived. The waiver list in scripts/go-file-length-waivers.txt is
for offenders that predate the cap and may only shrink; adding new growth to it
would retire the gate for the next file too.

The seam the gate asks for was already there. replydraft.go is the drafting CALL —
the prompt, the schema, the retry, the validation, the bounds — and the voice layer
is a profile laid over it: load the actor's active profile, run the draft under it,
check the result against the profile's avoid rules, record the outcome as a voice
signal. That whole layer moves to replydraftvoice.go, 126 lines, leaving 410 and 153.

voiceBlockFor stays behind deliberately: both system variants take it, so it belongs
with the request builder that reads it rather than with the profile that fills it.

The moved lines are re-gated by the strict linter (new-from-merge-base treats a move
as new), and they pass it unchanged — no reshaping was needed to get them through,
which is the outcome a pure move should have.
